youtube-automation/requirements.txt
google-generativeai>=0.3.0
youtube-transcript-api>=0.6.0
requests>=2.31.0
feedparser>=6.0.0
